+++ common/src/main/kotlin/com/cobblemon/mod/common/client/gui/interact/wheel/InteractWheelGUI.kt
@@ -11,6 +11,7 @@ package com.cobblemon.mod.common.client.gui.interact.wheel
 import com.cobblemon.mod.common.api.gui.blitk
 import com.cobblemon.mod.common.client.gui.CobblemonRenderable
 import com.cobblemon.mod.common.util.cobblemonResource
+import com.cobblemon.mod.common.util.isInventoryKeyPressed
 import com.google.common.collect.Multimap
 import net.minecraft.client.gui.GuiGraphics
 import net.minecraft.client.gui.screens.Screen
@@ -21,6 +22,14 @@ class InteractWheelGUI(private val options: Multimap<Orientation, InteractWheelO
     companion object {
         const val SIZE = 170
         private val backgroundResource = cobblemonResource("textures/gui/interact/interact_wheel_base.png")
+    }
+
+    override fun keyPressed(keyCode: Int, scanCode: Int, modifiers: Int): Boolean {
+        if (isInventoryKeyPressed(minecraft, keyCode, scanCode)) {
+            onClose()
+            return true
+        }
+        return super.keyPressed(keyCode, scanCode, modifiers)
     }
 
     private val buttons = mutableListOf<InteractWheelButton>()